Addressing Gal's comments
